+++ worker/src/handles/UdpSocketHandle.cpp
@@ -2,6 +2,9 @@
 // #define MS_LOG_DEV_LEVEL 3
 
 #include "handles/UdpSocketHandle.hpp"
+#ifdef MS_LIBURING_SUPPORTED
+#include "DepLibUring.hpp"
+#endif
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
 #include "Utils.hpp"
@@ -84,6 +87,15 @@ UdpSocketHandle::UdpSocketHandle(uv_udp_t* uvHandle) : uvHandle(uvHandle)
 
 		MS_THROW_ERROR("error setting local IP and port");
 	}
+
+#ifdef MS_LIBURING_SUPPORTED
+	err = uv_fileno(reinterpret_cast<uv_handle_t*>(this->uvHandle), std::addressof(this->fd));
+
+	if (err != 0)
+	{
+		MS_THROW_ERROR("uv_fileno() failed: %s", uv_strerror(err));
+	}
+#endif
 }
 
 UdpSocketHandle::~UdpSocketHandle()
@@ -157,6 +169,30 @@ void UdpSocketHandle::Send(
 		return;
 	}
 
+#ifdef MS_LIBURING_SUPPORTED
+	{
+		if (!DepLibUring::IsActive())
+		{
+			goto send_libuv;
+		}
+
+		// Prepare the data to be sent.
+		// NOTE: If all SQEs are currently in use or no UserData entry is available we'll
+		// fall back to libuv.
+		auto prepared = DepLibUring::PrepareSend(this->fd, data, len, addr, cb);
+
+		if (!prepared)
+		{
+			MS_DEBUG_DEV("cannot send via liburing, fallback to libuv");
+
+			goto send_libuv;
+		}
+
+		return;
+	}
+#endif
+
+send_libuv:
 	// First try uv_udp_try_send(). In case it can not directly send the datagram
 	// then build a uv_req_t and use uv_udp_send().